packing code into exe

+++ zip_assets.py
@@ -1,6 +1,7 @@
 # !python
 
 import os
+import getopt
 import sys
 import zipfile
 import subprocess
@@ -31,7 +32,7 @@ def translateCompileErrorMapping(originalString):
   return 'Parse error', msg
 
 
-def handleMoonFiles(zipOut, r, prefix, filePaths):
+def handleMoonFiles(zipOut, r, prefix, filePaths, stripCode):
   sourceMappings = ''
   for path in filePaths:
     luaFilePath = path.replace('.moon', '.lua')
@@ -49,12 +50,14 @@ def handleMoonFiles(zipOut, r, prefix, filePaths):
     else:
       mapping = subprocess.check_output([LUA_EXE, MOON_COMPILER, '-X', path])
       sourceMappings+='["%s"]={%s},'%(fileKey, translateMapping(mapping))
-    zipOut.write(luaFilePath, os.path.join("assets", prefix, rp))
-    os.remove(luaFilePath)
+
+    if not stripCode:
+      zipOut.write(luaFilePath, os.path.join("assets", prefix, rp))
+      os.remove(luaFilePath)
 
   return sourceMappings
 
-def zipDir(zipOut, r, prefix=""):
+def zipDir(zipOut, r, stripCode=False, prefix=""):
   convertScript = (subprocess.check_output([LUA_EXE, 'simple_moon_compile.lua', '-stdout', 'convert_flash_export.moon']))
 
   moonFiles = []
@@ -74,13 +77,13 @@ def zipDir(zipOut, r, prefix=""):
     for f in files:
       p = os.path.join(root, f)
       rp = os.path.relpath(p, r)
-      if not rp.startswith(".") and not 'moonscript' in p and p.endswith(('.lua','.moon', '.png', '.xml','.txt')):
+      if not rp.startswith(".") and not 'moonscript' in p and ((not stripCode and p.endswith('.lua')) or p.endswith(('.moon', '.png', '.xml','.txt'))):
         if p.endswith('.moon'):
           moonFiles.append(p)
         else:
           #print rp, '<-', p
           zipOut.write(p, os.path.join("assets", prefix, rp))
-  return handleMoonFiles(zipOut, r, prefix, moonFiles)
+  return handleMoonFiles(zipOut, r, prefix, moonFiles, stripCode)
 
 
 if __name__ == '__main__':
@@ -88,15 +91,20 @@ if __name__ == '__main__':
   appPath = os.path.join("..", "bounce")
   outZipPath = os.path.join('bin', 'assets.zip')
   frameworkSrcPath = os.path.join('src', 'lua')
-  offs = 0
-  if len(sys.argv) > 1 and sys.argv[1] == 'test':
-    offs = 1
-  if len(sys.argv) > 1+offs: 
-    appPath = sys.argv[1+offs]
-  if len(sys.argv) > 2+offs:
-    outZipPath = sys.argv[2+offs]
-  if len(sys.argv) > 3+offs:
-    frameworkSrcPath = sys.argv[3+offs]
+  
+  opts, args = getopt.getopt(sys.argv[1:],"ts")
+
+  stripCode = ('-s','') in opts
+
+  if stripCode:
+    print 'Stripping code'
+
+  if len(args)>0:
+    appPath = args[0]
+  if len(args)>1:
+    outZipPath = args[1]
+  if len(args)>2:
+    frameworkSrcPath = args[2]
 
   if not os.path.exists(appPath):
     raise Exception('Invalid path: %s'%appPath)
@@ -105,15 +113,17 @@ if __name__ == '__main__':
   print 'Zip path:', outZipPath
 
   zipOutFile = zipfile.ZipFile(outZipPath, "w")
-  moonSourceMappings = zipDir(zipOutFile, frameworkSrcPath, "framework")
-  if 'test' in sys.argv:
-    moonSourceMappings += zipDir(zipOutFile, "src/luatest", "framework/test")
-    moonSourceMappings += zipDir(zipOutFile, "src/luatest/data", "framework/testdata")
-  moonSourceMappings += zipDir(zipOutFile, appPath)
+  moonSourceMappings = zipDir(zipOutFile, frameworkSrcPath, stripCode, "framework")
+  if ('-t','') in opts:
+    print 'Including test sources'
+    moonSourceMappings += zipDir(zipOutFile, "src/luatest", "framework/test", stripCode)
+    moonSourceMappings += zipDir(zipOutFile, "src/luatest/data", "framework/testdata", stripCode)
+  moonSourceMappings += zipDir(zipOutFile, appPath, stripCode)
 
   with open('moon_source_mappings.lua', 'w') as out:
     out.write('return {%s}'%moonSourceMappings.replace("\\", "/"));
-  zipOutFile.write('moon_source_mappings.lua', os.path.join('assets', 'moon_source_mappings.lua'))
+  if not stripCode:
+    zipOutFile.write('moon_source_mappings.lua', os.path.join('assets', 'moon_source_mappings.lua'))
   os.remove('moon_source_mappings.lua')
 
   zipOutFile.close()
